refactoring: moved interfaces/consts to common part
