+++ src/cli/models/index.ts
@@ -7,4 +7,3 @@
  */
 
 export * from './cli-options';
-export * from './output-format';
