+++ src/common/dictionary.ts
