+++ src/common/index.ts
@@ -5,8 +5,8 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
  */
-import { Options } from '../../common';
 
-export interface CliOptions extends Options {
-  output?: string;
-}
+export * from './dictionary';
+export * from './directory';
+export * from './options';
+export * from './output-format';
