+++ src/common/options.ts
@@ -5,8 +5,13 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
  */
-import { Options } from '../../common';
 
-export interface CliOptions extends Options {
-  output?: string;
+export interface Options {
+  single: boolean;
+  svgo: boolean;
 }
+
+export const defaultOptions = {
+  single: false,
+  svgo: true,
+};
