+++ src/core/manifest-parser.ts
@@ -6,8 +6,9 @@
  * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
  */
 
-import { readFileSync }                                                             from 'fs';
-import { Dictionary, Directory, Manifest, ManifestFile, RawResource, ResourcesMap } from './models';
+import { readFileSync }                                      from 'fs';
+import { Dictionary, Directory }                             from '../common';
+import { Manifest, ManifestFile, RawResource, ResourcesMap } from './models';
 
 export const manifestInfo: Manifest = {
   artboards: [],
