+++ src/core/utils/gradients-list.ts
@@ -1,4 +1,4 @@
 import { XMLElementOrXMLNode } from 'xmlbuilder';
-import { Dictionary }          from '../models';
+import { Dictionary }          from '../../common';
 
 export const gradients: Dictionary<XMLElementOrXMLNode> = {};
